feat: Preloader завязать на isLoading
