+++ src/components/MoviesCardList/MoviesCardList.css
@@ -25,7 +25,7 @@
   justify-content: center;
   border-top: 1px solid #E8E8E8;
   margin: 0;
-  padding: 80px 0 30px;
+  padding: 80px 0 80px;
 }
 
 @media screen and (max-width: 1279px) {
@@ -42,7 +42,7 @@
     grid-template-rows: repeat(1, 235px);
     grid-auto-rows: 235px;
     grid-template-columns: repeat(1, 1fr);
-    padding: 50px 0 30px;
+    padding: 50px 0 80px;
     gap: 20px 0;
   }
 }
@@ -61,7 +61,7 @@
   align-content: center;
   width: 320px;
   height: 36px;
-  margin: 50px auto 80px;
+  margin: 0 auto 80px;
   padding: 0;
 }
 
@@ -69,4 +69,12 @@
   .movies-list__load {
     width: 240px;
   }
+}
+
+.movies-list__info {
+  text-align: center;
+  font-size: 16px;
+  font-weight: 500;
+  margin: 0;
+  padding: 20px 0 100px;
 }
